pipeline/requirements.txt
pylint
pytest
pandas
python-dotenv
requests
snowflake-connector-python